임시저장-옵션 별도의 파일로 분리
